Fork oslo.config.generator from stable/icehouse
Because we'll need to generate configuration files from external
libraries, we need to fork the generator from oslo.  So, for now
we just did a direct copy from the stable/icehouse branch and
removed the module from openstack-common.conf.

Once we update our libraries to stable/icehouse we can basically
revert this commit.

Change-Id: Ida3c86ced8f805aca9bea30d44a2f051547190c8
Signed-off-by: Paul Belanger <paul.belanger@polybeacon.com>